+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Internal refactoring to Angle Bracket components
 - Removed `is-searching` class name in favour of `is-busy` and/or aria attribute
 - Removed `@open` argument. (Please use the API instead)
+- Changed signature of `@onUpdate`, value is no longer passed as first argument
 
 ## 10.0.9